+++ BDSKOutlineView.m
@@ -159,7 +159,7 @@ static char BDSKOutlineViewFontDefaultsObservationContext;
     NSUserDefaults *sud = [NSUserDefaults standardUserDefaults];
     
     NSString *fontName = [sud objectForKey:fontNamePrefKey];
-    float fontSize = [sud floatForKey:fontSizePrefKey];
+    CGFloat fontSize = [sud floatForKey:fontSizePrefKey];
 	NSFont *font = nil;
         
     if(fontName != nil)
@@ -175,7 +175,7 @@ static char BDSKOutlineViewFontDefaultsObservationContext;
 - (void)outlineViewFontChanged {
     if (fontNamePreferenceKey && fontSizePreferenceKey) {
         NSString *fontName = [[NSUserDefaults standardUserDefaults] objectForKey:fontNamePreferenceKey];
-        float fontSize = [[NSUserDefaults standardUserDefaults] floatForKey:fontSizePreferenceKey];
+        CGFloat fontSize = [[NSUserDefaults standardUserDefaults] floatForKey:fontSizePreferenceKey];
         NSFont *font = nil;
         
         if(fontName != nil)
@@ -202,7 +202,7 @@ static char BDSKOutlineViewFontDefaultsObservationContext;
 - (void)updateFontPanel:(NSNotification *)notification {
     if ([[[self window] firstResponder] isEqual:self] && fontNamePreferenceKey && fontSizePreferenceKey) {
         NSString *fontName = [[NSUserDefaults standardUserDefaults] objectForKey:fontNamePreferenceKey];
-        float fontSize = [[NSUserDefaults standardUserDefaults] floatForKey:fontSizePreferenceKey];
+        CGFloat fontSize = [[NSUserDefaults standardUserDefaults] floatForKey:fontSizePreferenceKey];
         [[NSFontManager sharedFontManager] setSelectedFont:[NSFont fontWithName:fontName size:fontSize] isMultiple:NO];
     }
 }
@@ -529,7 +529,7 @@ static char BDSKOutlineViewFontDefaultsObservationContext;
     [self noteHeightOfRowsWithIndexesChanged:[NSIndexSet indexSetWithIndexesInRange:NSMakeRange(0, [self numberOfRows])]];
 }
 
-- (float)rowHeightForFont:(NSFont *)font {
+- (CGFloat)rowHeightForFont:(NSFont *)font {
     return [NSLayoutManager defaultViewLineHeightForFont:font];
 }
