+++ BDSKTypeInfoEditor.m
@@ -745,11 +745,11 @@ static BDSKTypeInfoEditor *sharedTypeInfoEditor;
 
 #pragma mark Splitview delegate methods
 
-- (float)splitView:(NSSplitView *)sender constrainMinCoordinate:(float)proposedMin ofSubviewAt:(NSInteger)offset{
+- (CGFloat)splitView:(NSSplitView *)sender constrainMinCoordinate:(CGFloat)proposedMin ofSubviewAt:(NSInteger)offset{
 	return proposedMin + 50.0;
 }
 
-- (float)splitView:(NSSplitView *)sender constrainMaxCoordinate:(float)proposedMax ofSubviewAt:(NSInteger)offset{
+- (CGFloat)splitView:(NSSplitView *)sender constrainMaxCoordinate:(CGFloat)proposedMax ofSubviewAt:(NSInteger)offset{
 	return proposedMax - 50.0;
 }
